was able to fix bug again for delete/update on all posts and converted switch statements to if statements.

+++ public/js/updatePost.js
@@ -7,71 +7,61 @@ const updatePost = async (e) => {
   const postId = e.target.getAttribute('data-id');
   const postAction = e.target.getAttribute('data-func');
 
-  switch (postAction) {
-    case 'update':
-      const updatedPost = {
-        'title': title,
-        'description': description
+  if(postAction === 'update') {
+    const updatedPost = {
+      'title': title,
+      'description': description
+    }
+    const updateResponse = await fetch(`/api/posts/${postId}`, {
+      method: 'PUT',
+      body: JSON.stringify(updatedPost),
+      headers: {
+        'Content-Type': 'application/json'
       }
+    });
 
-      const updateResponse = await fetch(`/api/posts/${postId}`, {
-        method: 'PUT',
-        body: JSON.stringify(updatedPost),
-        headers: {
-          'Content-Type': 'application/json'
-        }
-      });
-  
-      if(updateResponse.ok) {
-        document.location.replace('/dashboard');
-      } else {
-        alert('Failed to update post');
-      }
-
-    break;
-
-    default:
-      break;
+    if(updateResponse.ok) {
+      document.location.replace('/dashboard');
+    } else {
+      alert('Failed to update post');
+    }
+    return;
   }
+  return console.log("An error has occurred!");;
 }
 
 
 const updateDelPost = async (e) => {
   const reqAction = e.target.getAttribute('data-func');
 
-  switch (reqAction) {
-
-    case 'update':  
-      if (e.target.hasAttribute('data-id')) {
+  if(reqAction === 'update') {
+    if (e.target.hasAttribute('data-id')) {
       const id = e.target.getAttribute('data-id');
       const idNum = parseInt(id, 10);
       location.href=`/post/${idNum}`;
     }
-    break;
-  
-    case 'delete':
-      if (e.target.hasAttribute('data-id')) {
-        const id = e.target.getAttribute('data-id');
-        const idNum = parseInt(id, 10);
-  
-        const response = await fetch(`/api/posts/${idNum}`, {
-          method: 'DELETE',
-        });
-  
-        if (response.ok) {
-          document.location.reload();
-        } else {
-          alert('Failed to delete project');
-        }
-      }
-      break;
+    return;
+  }
+  if(reqAction === 'delete') {
+    if (e.target.hasAttribute('data-id')) {
+      const id = e.target.getAttribute('data-id');
+      const idNum = parseInt(id, 10);
 
-    default:
-      break;
+      const response = await fetch(`/api/posts/${idNum}`, {
+        method: 'DELETE',
+      });
+
+      if (response.ok) {
+        document.location.reload();
+      } else {
+        alert('Failed to delete project');
+      }
+    }
+    return;
   }
 }
 
-let updateDelBtn = document.querySelector('.posts-container');
+let updateDelBtn = document.querySelector('.posts-containers');
 
 if(updateDelBtn) {
   updateDelBtn.addEventListener('click', updateDelPost);
